+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - [sdk] Fix cancellation.
 - [sdk] Change method name from `cancelExpired` to `cancel`.
 - [sdk, router] Add `network` param for abstracting messaging configuration.
+- [sdk] Rename `startTransfer` and `finishTransfer` to `prepareTransfer` and `fulfillTransfer`.
 
 ## 0.0.15